DLCs_Main: Use function/data pointers from SADXModLoader if they're already defined.
Removed definitions for the following function pointers:
- sub_4B79C0() -> DisplayHintText()
- sub_49D6C0() -> DynamicCOL_AddFromEntity()

For DisplayHintText(), the first parameter no longer needs to be casted
to char*, since SADXModLoader has the correct type. (const char *const *)

Removed some variable pointer definitions. These were already defined in
SADXModLoader, but with different names. To avoid having to rename
everything, some macros were added to convert the names.

Famitsu.cpp: One of the DisplayHintText() calls used a duration of 140,
but all the other surrounding calls used 180. Changed it to 180, since
it was probably a typo.